Rename: themeApi -> getRecipients
